refactor(calendar): lift Suggestion type to types.ts to remove sidebar duplication
AdminListSidebar.tsx previously redeclared the Suggestion shape inline
with a comment claiming the duplication was needed to avoid pulling
server-only deps into the browser bundle. The real reason was the
.server.ts boundary: the project's ESLint rule forbids importing from
*.server.ts files in browser-bound code, even via \`import type\`.

types.ts has no .server suffix and is already the home of ParsedFeed /
ParsedVEvent / ParsedSeries — pure interfaces shared across the boundary.
Moving Suggestion there removes the duplication without touching the
guard rail. build.server.ts re-exports the type so external imports
keep working unchanged.
